+++ src/Namespaced/Tests/NamespacedCachePoolTest.php
@@ -12,6 +12,7 @@
 namespace Cache\Namespaced\Tests;
 
 use Cache\Namespaced\NamespacedCachePool;
+use PHPUnit\Framework\TestCase;
 use Psr\Cache\CacheItemInterface;
 
 /**
@@ -19,7 +20,7 @@ use Psr\Cache\CacheItemInterface;
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class NamespacedCachePoolTest extends \PHPUnit_Framework_TestCase
+class NamespacedCachePoolTest extends TestCase
 {
     /**
      * @return \PHPUnit_Framework_MockObject_MockObject
@@ -124,7 +125,7 @@ class NamespacedCachePoolTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($returnValue, $pool->save($item));
     }
 
-    public function testSaveDeffered()
+    public function testSaveDeferred()
     {
         $item        = $this->getMockBuilder(CacheItemInterface::class)->getMock();
         $namespace   = 'ns';