+++ src/Adapter/Chain/Changelog.md
@@ -4,6 +4,25 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.0.0
+
+### Added
+
+* The `CachePoolChain` does now implement `Cache\TagInterop\TaggableCacheItemPoolInterface`
+
+### Removed
+
+* Removed deprecated function `clearTags`
+
+## 0.5.1
+
+### Fixed
+
+* Fixed issue with generator
+* Make sure `getItems` always save values back to previous pools
+
+## 0.5.0
+
 ### Added
 
 * Support for the new `TaggableCacheItemPoolInterface`. 
@@ -16,4 +35,4 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## 0.3.1
 
-No changelog before this version
+* No changelog before this version