+++ src/Adapter/Common/Changelog.md
@@ -2,7 +2,25 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release. 
 
-## UNRELEASED
+## 1.1.0
+
+### Added 
+
+- Support for storing binary data
+
+### Fixed
+
+- Issue with one character variables
+
+### Changed
+
+- Tests are now extending `PHPUnit\Framework\TestCase` 
+
+## 1.0.0
+
+* No changes since 0.4.0.
+
+## 0.4.0
 
 ### Added
 
@@ -40,4 +58,4 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
  
 ## 0.2.0
  
-No changelog before this version
+* No changelog before this version