+++ src/Adapter/Doctrine/Changelog.md
@@ -4,9 +4,26 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.0.0
+
+* No changes since 0.6.0
+
+## 0.6.0
+
 ### Added
 
 * Support for the new `TaggableCacheItemPoolInterface`. 
+* Support for PSR-16 SimpleCache
+
+### Changed
+
+* The behavior of `CacheItem::getTags()` has changed. It will not return the tags stored in the cache storage. 
+
+### Removed
+
+* `CacheItem::getExpirationDate()`. Use `CacheItem::getExpirationTimestamp()`
+* `CacheItem::getTags()`. Use `CacheItem::getPreviousTags()`
+* `CacheItem::addTag()`. Use `CacheItem::setTags()`
 
 ## 0.5.1
 
@@ -16,4 +33,4 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## 0.5.0
 
-No changelog before this version
+* No changelog before this version