+++ src/Adapter/Redis/Changelog.md
@@ -4,9 +4,33 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.1.0
+
+### Added
+* Allow \RedisArray and \RedisCluster objects for RedisCachePool
+
+## 1.0.0
+
+### Fixed
+
+* Fixed `$path` variable not initialized in `clearOneObjectFromCache`.
+
+## 0.5.0
+
 ### Added
 
 * Support for the new `TaggableCacheItemPoolInterface`. 
+* Support for PSR-16 SimpleCache
+
+### Changed
+
+* The behavior of `CacheItem::getTags()` has changed. It will not return the tags stored in the cache storage. 
+
+### Removed
+
+* `CacheItem::getExpirationDate()`. Use `CacheItem::getExpirationTimestamp()`
+* `CacheItem::getTags()`. Use `CacheItem::getPreviousTags()`
+* `CacheItem::addTag()`. Use `CacheItem::setTags()`
 
 ## 0.4.2
 
@@ -17,4 +41,4 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## 0.4.1
 
-No changelog before this version
+* No changelog before this version