+++ src/Adapter/Void/Changelog.md
@@ -4,16 +4,39 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.0.0
+
+* No changes since 0.4.1.
+
+## 0.4.1
+
+### Changed
+
+* We now support cache/hierarchical-cache: ^0.4
+
+## 0.4.0
+
 ### Added
 
 * Support for the new `TaggableCacheItemPoolInterface`. 
+* Support for PSR-16 SimpleCache
+
+### Changed
+
+* The behavior of `CacheItem::getTags()` has changed. It will not return the tags stored in the cache storage. 
+
+### Removed
+
+* `CacheItem::getExpirationDate()`. Use `CacheItem::getExpirationTimestamp()`
+* `CacheItem::getTags()`. Use `CacheItem::getPreviousTags()`
+* `CacheItem::addTag()`. Use `CacheItem::setTags()`
 
 ## 0.3.1
 
 ### Changed
 
-Updated dependencies
+* Updated dependencies
 
 ## 0.3.0
 
-No changelog before this version
+* No changelog before this version