+++ src/Bridge/SimpleCache/Changelog.md
@@ -4,6 +4,16 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.0.0
+
+* No changes since 0.1.1
+
+## 0.1.1
+
+### Fixed
+
+* Bugs with iterators
+
 ## 0.1.0
 
-First release
+* First release