+++ src/Hierarchy/Changelog.md
@@ -4,16 +4,26 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## UNRELEASED
 
+## 1.0.0
+
+* No changes since 0.4.0
+
+## 0.4.0
+
 ### Changed
 
 * `HierarchicalCachePoolTrait::getValueFormStore` was renamed to `HierarchicalCachePoolTrait::getDirectValue`
+
+### Removed
+
+* Dependency to `cache/taggable-cache`. 
 
 ## 0.3.0
 
 ### Changed
 
-The `HierarchicalPoolInterface` extends `CacheItemPoolInterface`
+* The `HierarchicalPoolInterface` extends `CacheItemPoolInterface`
 
 ## 0.2.1
 
-No changelog before this version
+* No changelog before this version