+++ src/Namespaced/Changelog.md
@@ -1,8 +1,26 @@
 # Change Log
 
-The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release. 
+The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release.
 
 ## UNRELEASED
+
+## 1.0.0
+
+### Added
+
+* NamespacedCachePool implements HierarchicalPoolInterface
+
+## 0.1.3
+
+### Changed
+
+* Updated dependencies
+
+## 0.1.2
+
+### Fixed
+
+* Typos, documentation and general package improvements.
 
 ## 0.1.1
 
@@ -13,4 +31,4 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 
 ## 0.1.0
 
-First release
+* First release